% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ts.R
\name{check_ts}
\alias{check_ts}
\title{Check input time series.}
\usage{

}
\arguments{
\item{approx_spike}{}
}
\value{
TODO
}
\description{
Function for checking the validity of an input time series for further processing.
}
\details{
TODO
}
\author{
Sandro Groth
}
