% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{crop_fast}
\alias{crop_fast}
\title{Fast Raster Cropping}
\usage{
crop_fast(in_file, out_file, aoi_ext, driver = "GTiff", dtype = "INT2U")
}
\description{
TODO
}
\details{
TODO
}
\references{
Jakob Schwalb-Willmann and Henrik Fisser (2020). getSpatialData:
Get different kinds of freely available spatial datasets. R package version 0.1.0.
http://www.github.com/16eagle/getSpatialData/
}
