% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{divide_seasons}
\alias{divide_seasons}
\title{TODO: weighted harmonic modeling}
\usage{
divide_seasons(y, d, w, s_per_year = 1, amplitude_cutoff = 2000)
}
\arguments{
\item{s_per_year}{}
}
\description{
TODO: weighted harmonic modeling
}
