% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{extract_hdf}
\alias{extract_hdf}
\title{Extract subdatasets from .hdf archive}
\usage{
extract_hdf(hdf, sd, d_type = "INT2S")
}
\arguments{
\item{hdf}{Required character. Full path of hdf archive.}

\item{sd}{Required character. Specifies the subdataset that should be extracted.}
}
\description{
Extract subdatasets from .hdf archive
}
