% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{get_real_dates}
\alias{get_real_dates}
\title{Get real Date from DOY vector.

Applicable for products: MOD13Q1, MYD13Q1.}
\usage{
get_real_dates(d, comp_d)
}
\arguments{
\item{comp_d}{}
}
\description{
TODO
}
\details{
TODO
}
