% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{mes_uncertainties}
\alias{mes_uncertainties}
\title{Weights to measurement uncertainties conversion.}
\usage{
mes_uncertainties(w)
}
\arguments{
\item{w}{Numeric vector. A set of weights.}
}
\value{
Numeric vector. Measurement uncertainties.
}
\description{
Weights to measurement uncertainties conversion.
}
\details{
This function converts a set of weights, initially determined from the QA parameter of a
timeseries or estimated using an outlier detection method, into measurement uncertainties using following
transformation:
\deqn{\sigma=\frac{1}{w + 0.0001}}
}
\examples{
w <- c(0, 0.5, 1)
mes_uncertainties(w)

}
\references{
Eklundh, L., and Jönsson, P., 2017, TIMESAT 3.3 with seasonal trend decomposition and parallel
processing - Software Manual. Lund University, 92 pp.
}
\author{
Sandro Groth
}
