% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaWeights.R
\name{MODIS_summary_qa}
\alias{MODIS_summary_qa}
\title{Get weights based on MODIS summary QA values}
\usage{
MODIS_summary_qa(qa_val, w_min = 0, w_med = 0.5, w_max = 1)
}
\description{
Get weights based on MODIS summary QA values
}
