% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{mosaic_tiles}
\alias{mosaic_tiles}
\title{Mosaic tiles with same acquisition date.}
\usage{
mosaic_tiles(
  date_str,
  out_file,
  search_dir,
  driver = "GTiff",
  dtype = "SInt16"
)
}
\arguments{
\item{dtype}{}
}
\description{
TODO
}
\details{
TODO
}
