% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_MODIS.R
\name{prepare_MODIS}
\alias{prepare_MODIS}
\title{Prepare MODIS}
\usage{
prepare_MODIS(
  in_dir,
  out_dir,
  aoi,
  vi = "NDVI",
  product_name = NA,
  out_proj = NA,
  use_DOY = TRUE,
  use_QA = TRUE,
  keep_orig = TRUE,
  dtype = "INT2S",
  overwrite = TRUE,
  do_par = TRUE,
  cores = NA,
  progress = TRUE
)
}
\arguments{
\item{in_dir}{TODO}

\item{out_dir}{TODO}

\item{aoi}{TODO}

\item{vi}{TODO}

\item{product_name}{TODO}

\item{out_proj}{TODO}

\item{use_DOY}{TODO}

\item{use_QA}{TODO}

\item{keep_orig}{TODO}

\item{dtype}{TODO}

\item{overwrite}{TODO}

\item{do_par}{TODO}

\item{cores}{TODO}

\item{progress}{TODO}
}
\value{
TODO
}
\description{
\code{prepare_MODIS} prepares MODIS .hdf archives obtained by \link[phenoRS]{download_MODIS}
for curve fitting.
}
\details{
TODO
}
\examples{
\dontrun{TODO}

}
\seealso{
TODO
}
\author{
Sandro Groth
}
