% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMODIS.R
\name{prepMODIS}
\alias{prepMODIS}
\title{Prepare donwloaded MODIS archives for fitting.}
\usage{
prepMODIS(
  in_dir,
  out_dir,
  aoi,
  vi = "NDVI",
  out_proj = NA,
  correct_doy = TRUE,
  cores = NA
)
}
\description{
Prepare donwloaded MODIS archives for fitting.
}
