% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_IO.R
\name{reproject}
\alias{reproject}
\title{Image Reprojection}
\usage{
reproject(
  in_file,
  out_file,
  out_proj,
  driver = "GTiff",
  dtype = "SInt16",
  gdal = TRUE,
  overwrite = TRUE
)
}
\description{
TODO
}
\details{
TODO
}
