% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{spike_median}
\alias{spike_median}
\title{Modified Hampel Median Filter.}
\usage{
spike_median(v, w, ypts, w_min = 0, spk = 2, cpp = T)
}
\arguments{
\item{v}{numeric vector. Vegetation index values.}

\item{w}{numeric vector. Initial weights.}

\item{ypts}{numeric. The number of values per year. Used to determine the
moving window size.}

\item{w_min}{numeric. Specifies the weight that should be assigned to outliers.}

\item{spk}{numeric. The Spike parameter. the higher, the more values are classified
outliers.}
}
\description{
Modified Hampel Median Filter.
}
\details{
#TODO
}
\examples{
## Create a random normal distribution with outliers.
values <- c(rnorm(19), 50, rnorm(19), -50, rnorm(20))

## set all initial weights to 1
init_weights <- rep(1, 60)

## get the spike filtered weights
filtered_weights <- spike_median(values, init_weights, 20, 0, 2)

}
\references{
Hampel F. R., ”The influence curve and its role in robust estimation,”
            Journal of the American Statistical Association, 69, 382–393, 1974

            Eklundh, L., and Jönsson, P., 2017, TIMESAT 3.3 with seasonal trend
            decomposition and parallel processing - Software Manual.
            Lund University, 92 pp.
}
\author{
Sandro Groth
}
