% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ts.R
\name{check_ts}
\alias{check_ts}
\title{Check input time series.}
\usage{
check_ts(
  y,
  d,
  w,
  valid_min,
  amplitude_cutoff = 200,
  y_min = 200,
  y_max = 10000,
  w_min = 0,
  approx_spike = TRUE
)
}
\arguments{
\item{approx_spike}{}
}
\description{
Check input time series.
}
