% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction.R
\name{correct_doy}
\alias{correct_doy}
\title{Correct DOY layer of extracted MODIS raster objects.}
\usage{
correct_doy(r_obj, comp_year)
}
\arguments{
\item{r_obj}{Required raster object. Holds MODIS DOY as integers.}

\item{comp_year}{Required integer. Year of the MODIS composite.}
}
\description{
TODO
}
